backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# LangChain & AI (Compatible versions)
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-community==0.3.5

# Google Gemini
google-generativeai==0.8.3

# Database
supabase==2.9.0
psycopg2-binary==2.9.10
pgvector==0.3.5

# Pydantic (Compatible with LangChain)
pydantic==2.9.2
pydantic-settings==2.6.1

# Utilities
numpy==1.26.4
httpx==0.27.2

# PDF Generation
reportlab==4.2.5
PyPDF2==3.0.1